* .editorconfig updates
* Uniform WaitForOverlay... methods.
* Setting for killing the tested app after a failed test.
* Remove Json stuff from the labels template.

+++ main/source/UiCommonFunctions/Base/TcBaseTestClass.cs
@@ -289,7 +289,7 @@ namespace HomeZone.UiCommonFunctions.Base
         /// </summary>
         protected virtual void DoTestCleanup()
         {
-            if( TestContext.CurrentTestOutcome == UnitTestOutcome.Failed )
+            if( TestContext.CurrentTestOutcome == UnitTestOutcome.Failed && TestSettings.KillTestedAppAfterFailedTest )
             {
                 mTestedAppProcess?.Kill();
             }
